Reorganizing the sorting javascript files
