+++ static/js/simpletreemenu.js
@@ -7,25 +7,53 @@ ddtreemenu.openfolder = "/img/open.gif"; // Set image path to "open" folder imag
 ////////// No need to edit beyond here ///////////////////////////
 
 /**
- * Creates a tree menu with optional persistence.
+ * Expands a UL element and any of its parent ULs.
  * @param {string} treeid - The ID of the tree element.
- * @param {boolean} enablepersist - Whether to enable persistence.
- * @param {number} persistdays - Number of days to persist the state.
+ * @param {HTMLElement} ulelement - The UL element to expand.
  */
-ddtreemenu.createTree = function (treeid, enablepersist, persistdays) {
-    let ultags = document.getElementById(treeid).getElementsByTagName("ul");
-    if (typeof persisteduls[treeid] === "undefined") {
-        persisteduls[treeid] = (enablepersist === true && ddtreemenu.getCookie(treeid) !== "") ? ddtreemenu.getCookie(treeid).split(",") : "";
+ddtreemenu.expandSubTree = function (treeid, ulelement) {
+    let rootnode = document.getElementById(treeid);
+    let currentnode = ulelement;
+    currentnode.style.display = "block";
+    currentnode.parentNode.style.backgroundImage = "url(" + ddtreemenu.openfolder + ")";
+    while (currentnode !== rootnode) {
+        if (currentnode.tagName === "UL") {
+            currentnode.style.display = "block";
+            currentnode.setAttribute("rel", "open");
+            currentnode.parentNode.style.backgroundImage = "url(" + ddtreemenu.openfolder + ")";
+        }
+        currentnode = currentnode.parentNode;
     }
-    for (let i = 0; i < ultags.length; i++) {
-        ddtreemenu.buildSubTree(treeid, ultags[i], i);
+};
+
+/**
+ * Prevents an action from bubbling upwards.
+ * @param {Event} e - The event to prevent propagation.
+ */
+ddtreemenu.preventpropagate = function (e) {
+    if (typeof e !== "undefined") {
+        e.stopPropagation();
+    } else {
+        event.cancelBubble = true;
     }
-    if (enablepersist === true) {
-        let durationdays = (typeof persistdays === "undefined") ? 1 : parseInt(persistdays);
-        ddtreemenu.dotask(window, function () {
-            ddtreemenu.rememberstate(treeid, durationdays);
-        }, "unload"); // Save opened UL indexes on body unload
+};
+
+/**
+ * Searches an array for the entered value. If found, deletes the value from the array.
+ * @param {Array} thearray - The array to search.
+ * @param {any} value - The value to search for.
+ * @returns {boolean} - True if the value is found, false otherwise.
+ */
+ddtreemenu.searcharray = function (thearray, value) {
+    let isfound = false;
+    for (let i = 0; i < thearray.length; i++) {
+        if (thearray[i] === value) {
+            isfound = true;
+            thearray.shift();
+            break;
+        }
     }
+    return isfound;
 };
 
 /**
@@ -68,38 +96,44 @@ ddtreemenu.buildSubTree = function (treeid, ulelement, index) {
 };
 
 /**
- * Expands a UL element and any of its parent ULs.
- * @param {string} treeid - The ID of the tree element.
- * @param {HTMLElement} ulelement - The UL element to expand.
+ * Assigns a function to execute to an event handler.
+ * @param {Window} target - The target window.
+ * @param {Function} functionref - The function to execute.
+ * @param {string} tasktype - The type of event (e.g., "unload").
  */
-ddtreemenu.expandSubTree = function (treeid, ulelement) {
-    let rootnode = document.getElementById(treeid);
-    let currentnode = ulelement;
-    currentnode.style.display = "block";
-    currentnode.parentNode.style.backgroundImage = "url(" + ddtreemenu.openfolder + ")";
-    while (currentnode !== rootnode) {
-        if (currentnode.tagName === "UL") {
-            currentnode.style.display = "block";
-            currentnode.setAttribute("rel", "open");
-            currentnode.parentNode.style.backgroundImage = "url(" + ddtreemenu.openfolder + ")";
-        }
-        currentnode = currentnode.parentNode;
+ddtreemenu.dotask = function (target, functionref, tasktype) {
+    let eventType = (window.addEventListener) ? tasktype : "on" + tasktype;
+    if (target.addEventListener) {
+        target.addEventListener(eventType, functionref, false);
+    } else if (target.attachEvent) {
+        target.attachEvent(eventType, functionref);
     }
 };
 
 /**
- * Expands or contracts all UL elements.
- * @param {string} treeid - The ID of the tree element.
- * @param {string} action - The action to perform ("expand" or "contract").
+ * Gets the value of a cookie.
+ * @param {string} Name - The name of the cookie.
+ * @returns {string} - The value of the cookie.
  */
-ddtreemenu.flatten = function (treeid, action) {
-    let ultags = document.getElementById(treeid).getElementsByTagName("ul");
-    for (let i = 0; i < ultags.length; i++) {
-        ultags[i].style.display = (action === "expand") ? "block" : "none";
-        let relvalue = (action === "expand") ? "open" : "closed";
-        ultags[i].setAttribute("rel", relvalue);
-        ultags[i].parentNode.style.backgroundImage = (action === "expand") ? "url(" + ddtreemenu.openfolder + ")" : "url(" + ddtreemenu.closefolder + ")";
+ddtreemenu.getCookie = function (Name) {
+    let re = new RegExp(Name + "=[^;]+", "i");
+    if (document.cookie.match(re)) {
+        return document.cookie.match(re)[0].split("=")[1];
     }
+    return "";
+};
+
+
+/**
+ * Sets the value of a cookie.
+ * @param {string} name - The name of the cookie.
+ * @param {string} value - The value of the cookie.
+ * @param {number} days - Number of days to persist the cookie.
+ */
+ddtreemenu.setCookie = function (name, value, days) {
+    let expireDate = new Date();
+    let expstring = expireDate.setDate(expireDate.getDate() + parseInt(days));
+    document.cookie = name + "=" + value + "; expires=" + expireDate.toGMTString() + "; path=/";
 };
 
 /**
@@ -122,72 +156,39 @@ ddtreemenu.rememberstate = function (treeid, durationdays) {
 };
 
 /**
- * Gets the value of a cookie.
- * @param {string} Name - The name of the cookie.
- * @returns {string} - The value of the cookie.
+ * Creates a tree menu with optional persistence.
+ * @param {string} treeid - The ID of the tree element.
+ * @param {boolean} enablepersist - Whether to enable persistence.
+ * @param {number} persistdays - Number of days to persist the state.
  */
-ddtreemenu.getCookie = function (Name) {
-    let re = new RegExp(Name + "=[^;]+", "i");
-    if (document.cookie.match(re)) {
-        return document.cookie.match(re)[0].split("=")[1];
+ddtreemenu.createTree = function (treeid, enablepersist, persistdays) {
+    let ultags = document.getElementById(treeid).getElementsByTagName("ul");
+    if (typeof persisteduls[treeid] === "undefined") {
+        persisteduls[treeid] = (enablepersist === true && ddtreemenu.getCookie(treeid) !== "") ? ddtreemenu.getCookie(treeid).split(",") : "";
     }
-    return "";
-};
-
-/**
- * Sets the value of a cookie.
- * @param {string} name - The name of the cookie.
- * @param {string} value - The value of the cookie.
- * @param {number} days - Number of days to persist the cookie.
- */
-ddtreemenu.setCookie = function (name, value, days) {
-    let expireDate = new Date();
-    let expstring = expireDate.setDate(expireDate.getDate() + parseInt(days));
-    document.cookie = name + "=" + value + "; expires=" + expireDate.toGMTString() + "; path=/";
-};
-
-/**
- * Searches an array for the entered value. If found, deletes the value from the array.
- * @param {Array} thearray - The array to search.
- * @param {any} value - The value to search for.
- * @returns {boolean} - True if the value is found, false otherwise.
- */
-ddtreemenu.searcharray = function (thearray, value) {
-    let isfound = false;
-    for (let i = 0; i < thearray.length; i++) {
-        if (thearray[i] === value) {
-            isfound = true;
-            thearray.shift();
-            break;
-        }
+    for (let i = 0; i < ultags.length; i++) {
+        ddtreemenu.buildSubTree(treeid, ultags[i], i);
     }
-    return isfound;
-};
-
-/**
- * Prevents an action from bubbling upwards.
- * @param {Event} e - The event to prevent propagation.
- */
-ddtreemenu.preventpropagate = function (e) {
-    if (typeof e !== "undefined") {
-        e.stopPropagation();
-    } else {
-        event.cancelBubble = true;
+    if (enablepersist === true) {
+        let durationdays = (typeof persistdays === "undefined") ? 1 : parseInt(persistdays);
+        ddtreemenu.dotask(window, function () {
+            ddtreemenu.rememberstate(treeid, durationdays);
+        }, "unload"); // Save opened UL indexes on body unload
     }
 };
 
 /**
- * Assigns a function to execute to an event handler.
- * @param {Window} target - The target window.
- * @param {Function} functionref - The function to execute.
- * @param {string} tasktype - The type of event (e.g., "unload").
+ * Expands or contracts all UL elements.
+ * @param {string} treeid - The ID of the tree element.
+ * @param {string} action - The action to perform ("expand" or "contract").
  */
-ddtreemenu.dotask = function (target, functionref, tasktype) {
-    let eventType = (window.addEventListener) ? tasktype : "on" + tasktype;
-    if (target.addEventListener) {
-        target.addEventListener(eventType, functionref, false);
-    } else if (target.attachEvent) {
-        target.attachEvent(eventType, functionref);
+ddtreemenu.flatten = function (treeid, action) {
+    let ultags = document.getElementById(treeid).getElementsByTagName("ul");
+    for (let i = 0; i < ultags.length; i++) {
+        ultags[i].style.display = (action === "expand") ? "block" : "none";
+        let relvalue = (action === "expand") ? "open" : "closed";
+        ultags[i].setAttribute("rel", relvalue);
+        ultags[i].parentNode.style.backgroundImage = (action === "expand") ? "url(" + ddtreemenu.openfolder + ")" : "url(" + ddtreemenu.closefolder + ")";
     }
 };
