+++ templates/sorting.html
@@ -20,9 +20,11 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/d3/3.5.17/d3.min.js" integrity="sha384-N8EP0Yml0jN7e0DcXlZ6rt+iqKU9Ck6f1ZQ+j2puxatnBq4k9E8Q6vqBcY34LNbn" crossorigin="anonymous"></script>
 
     <script type="text/javascript" src="{{ url_for('static', filename='js/dragAndDropTree.js') }}"></script>
-    <script type="text/javascript" src="{{ url_for('static', filename='js/difftree.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('static', filename='js/sorting/difftree.js') }}"></script>
 
-    <script type="text/javascript" src="{{ url_for('static', filename='js/sorting.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('static', filename='js/sorting/captchaOperations.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('static', filename='js/sorting/formOperations.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('static', filename='js/sorting/main.js') }}"></script>
 
     <script type="text/javascript" >
         {% for x in filelist %}
